requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.1.13
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cors-headers==4.3.1
django-environ==0.10.0
django-timezone-field==7.0
djangorestframework==3.15.1
flower==2.0.1
gunicorn==22.0.0
humanize==4.10.0
idna==3.7
kombu==5.3.7
packaging==24.1
pip==24.0
psycopg2-binary==2.9.5
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==4.6.0
requests==2.32.3
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
wheel==0.43.0
django-debug-toolbar==4.3.0